+++ TerminalGameEngine/Obstacle.cpp
@@ -2,7 +2,7 @@
 #include "AudioManager.h"
 #include "TimeHelper.h"
 
-Obstacle::Obstacle(int xPos, int yPos, Direction moveDir, float moveSpeed)
+Obstacle::Obstacle(int xPos, int yPos, Direction moveDir, double moveSpeed)
     : MovingStraightObject(xPos, yPos, moveDir, moveSpeed) { }
 
 void Obstacle::OnCollisionEnter(GameObject* other, Direction collisionDirection)
